Revert "- Refactor listing management to support categories"
